convert files test (intentionally broken for now)
